backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-multipart
python-dotenv
agno
pydantic-settings
